+++ src/interpreter/cps_interpreter.rs
@@ -1,4 +1,4 @@
-use parser::*;
+use crate::reader::parser::*;
 
 use std::fmt;
 use std::collections::HashMap;
