+++ src/lib.rs
@@ -1,0 +1,14 @@
+pub mod reader;
+pub mod core;
+pub mod interpreter;
+
+//pub mod lexer;
+//mod parser;
+//mod interpreter;
+//mod ast_walk_interpreter;
+//mod cps_interpreter;
+
+
+
+//#[cfg(not(test))]
+//mod repl;
